EO-737 Added friendly label for Mailbox Providers in signals table
