+++ src/constants/index.js
@@ -139,3 +139,23 @@ export const ROLES = {
 };
 
 export const SUBACCOUNT_ROLES = [ROLES.SUBACCOUNT_REPORTING];
+
+export const MB_PROVIDERS = {
+  apple: 'Apple',
+  gmail: 'Gmail',
+  google_apps: 'Google Apps',
+  microsoft: 'Microsoft',
+  verizon_media: 'Verizon Media',
+  yahoo_uk: 'Yahoo UK',
+  us_broadband: 'US Broadband',
+  uk_providers: 'UK Providers',
+  italian_providers: 'Italian Providers',
+  russian_providers: 'Russian Providers',
+  french_providers: 'French Providers',
+  chinese_providers: 'Chinese Providers',
+  german_providers: 'German Providers',
+  canandian_providers: 'Canadian Providers',
+  japanese_providers: 'Japanese Providers',
+  other_european_providers: 'Other European Providers',
+  other: 'Other Providers'
+};
